clearly orgonizes dto by files and dirs
